+++ src/api/index.ts
@@ -1,81 +1,100 @@
-type InitOpts = { host?: string; check?: (path: string) => string };
+const options: {
+	/** Base url to prefix urls in fetch requests  */
+	host?: string;
+	/**
+	 * Intercepts url before getting passed to fetch
+	 * @param path url received from all api methods
+	 */
+	intercept?: (path: string) => string;
+	fetch?: typeof fetch;
+} = {};
+
+type SendOptions = { fetch: typeof fetch };
 type SendParams = { method: string; path: string; data?: BodyInit; token?: string };
-type MaussFetch = (url: string, init?: RequestInit) => Promise<Response>;
-
-const tryImport = async (browser: boolean) => {
-	if (browser) return window.fetch;
-	try {
-		const module = await import('node-fetch');
-		return <MaussFetch>(module.default as any);
-	} catch (err) {
-		console.warn(`Cannot use in server, "node-fetch" not available`);
-		return '';
-	}
-};
-
-const options: InitOpts & { fetch?: MaussFetch } = {};
-async function send({ method, path, data, token }: SendParams) {
+type SendOutput = Promise<{ response: Response; body: string | Record<string, any> }>;
+async function send({ fetch }: SendOptions, { method, path, data, token }: SendParams): SendOutput {
 	const browser = typeof window !== 'undefined';
-	if (!options.fetch) {
-		const module = await tryImport(browser);
-		if (typeof module !== 'string') options.fetch = module;
-		else if (!browser) return module;
-	}
 
 	const opts: RequestInit = { method };
-	opts.headers = {}; // Workaround TypeScript not correctly implying if put above
-	if (browser && typeof FormData !== 'undefined' && data instanceof FormData) {
-		opts.body = data;
-	} else if (data) {
-		opts.headers['Content-Type'] = 'application/json';
-		opts.body = JSON.stringify(data);
+	opts.headers = {}; // TS workaround for "Object is possibly 'undefined'."
+	if (data) {
+		if (
+			(typeof Blob !== 'undefined' && data instanceof Blob) ||
+			(typeof FormData !== 'undefined' && data instanceof FormData)
+		) {
+			opts.body = data;
+		} else if (typeof data === 'object') {
+			opts.headers['Content-Type'] = 'application/json';
+			opts.body = JSON.stringify(data);
+		} else if (typeof data === 'string') {
+			opts.headers['Content-Type'] = 'text/plain';
+			opts.headers['Content-Length'] = `${data.length}`;
+			opts.body = data;
+		}
 	}
 
 	if (token) {
 		opts.headers['Authorization'] = `Bearer ${token}`;
 	}
 
-	const { host: base, check } = options;
-	const fetch: MaussFetch = browser ? (window.fetch as any) : options.fetch;
+	const { host: base, intercept } = options;
 	/**
-	 * 1: check function precedes everything
+	 * 1: intercept function precedes everything
 	 * 2: use native fetch functionality w/o base
 	 * 3: force base domain for server side fetch
 	 */
 	const url =
-		(check && check(path)) ||
+		(intercept && intercept(path)) ||
 		(browser && (base ? `${base}/${path}` : path)) ||
 		`${base || 'localhost:3000'}/${path}`;
 
 	const res = await fetch(url, opts);
-	const text = await res.text();
-	try {
-		return JSON.parse(text);
-	} catch (err) {
-		return text;
-	}
+	const body = await (res.ok ? res.json() : res.text());
+	return { response: res, body };
 }
 
-export const init = async ({ host, check }: InitOpts) => {
-	const browser = typeof window !== 'undefined';
-	options.host = host && host.startsWith('http') ? host : undefined;
-	options.check = typeof check === 'function' ? check : undefined;
-	const module = await tryImport(browser);
-	if (typeof module !== 'string') options.fetch = module;
+function relay() {
+	if (typeof window !== 'undefined') return window.fetch;
+	if (typeof fetch !== 'undefined') return fetch;
+	if (typeof options.fetch !== 'undefined') return options.fetch;
+	throw new Error('No fetch provided. Use browser only or pass in a fetch function');
+}
+
+type MethodInit = string | { fetch?: typeof fetch; path: string };
+/**	GET request with fetch */
+export const get = async (init: MethodInit, token?: string) => {
+	const { fetch = relay(), path } = typeof init !== 'string' ? init : { path: init };
+	return await send({ fetch }, { method: 'GET', path, token });
+};
+/**	DELETE request with fetch */
+export const del = async (init: MethodInit, token?: string) => {
+	const { fetch = relay(), path } = typeof init !== 'string' ? init : { path: init };
+	return await send({ fetch }, { method: 'DELETE', path, token });
+};
+/**	POST request with fetch */
+export const post = async (init: MethodInit, data: any, token?: string) => {
+	const { fetch = relay(), path } = typeof init !== 'string' ? init : { path: init };
+	return await send({ fetch }, { method: 'POST', path, data, token });
+};
+/**	PUT request with fetch */
+export const put = async (init: MethodInit, data: any, token?: string) => {
+	const { fetch = relay(), path } = typeof init !== 'string' ? init : { path: init };
+	return await send({ fetch }, { method: 'PUT', path, data, token });
 };
 
-export function get(path: string, token?: string): Promise<Response> {
-	return send({ method: 'GET', path, token });
-}
-
-export function del(path: string, token?: string): Promise<Response> {
-	return send({ method: 'DELETE', path, token });
-}
-
-export function post(path: string, data: any, token?: string): Promise<Response> {
-	return send({ method: 'POST', path, data, token });
-}
-
-export function put(path: string, data: any, token?: string): Promise<Response> {
-	return send({ method: 'PUT', path, data, token });
-}
+export default {
+	/** Use api with additional options by initializing this function */
+	init({ host, intercept, fetch }: typeof options) {
+		options.host = host && /^https?/.test(host) ? host : '';
+		options.intercept = typeof intercept === 'function' ? intercept : undefined;
+		try {
+			options.fetch = fetch || relay();
+		} catch (error) {
+			if (typeof window === 'undefined') console.warn(error);
+		}
+	},
+	get,
+	del,
+	post,
+	put,
+};