+++ src/index.ts
@@ -1,2 +1,3 @@
-export { debounce, throttle } from './core/balancer';
-export { compare, comparator } from './core/inspector';
+export * from './core';
+export * from './core/balancer';
+export * from './core/inspector';