+++ src/math/index.ts
@@ -1,1 +1,3 @@
-export { permutation } from './set';
+import { permutation } from './set';
+
+export { permutation };