+++ src/typings/enhancer.ts
@@ -1,3 +1,0 @@
-export type Typify<T> = {
-	[K in keyof T]: Typify<T[K]>;
-};