+++ src/typings/helper.ts
@@ -1,1 +1,0 @@
-export type Overwrite<T, K> = Omit<T, keyof K> & K;